backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.9
pydantic==2.10.5
pydantic-settings==2.7.1
python-dotenv==1.0.1

litellm==1.61.7

PyMuPDF==1.25.2
Pillow==10.4.0

SQLAlchemy==2.0.37
aiosqlite==0.20.0
